4. Django/class-notes/022_DRF_CB_Views/requirements.txt
asgiref==3.5.0
black==21.12b0
click==8.0.3
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
zipp==3.8.0
